requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev,all --output-file=requirements_dev.txt pyproject.toml
#
asteval==1.0.2
    # via lmfit
black==24.8.0
    # via
    #   pytest-black
    #   qec-util (pyproject.toml)
build==1.2.1
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
cloudpickle==3.0.0
    # via scalene
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dem-decoders==0.1.1
    # via qec-util (pyproject.toml)
dill==0.3.8
    # via lmfit
fonttools==4.58.4
    # via matplotlib
galois==0.4.9
    # via qec-util (pyproject.toml)
gurobipy==12.0.2
    # via
    #   qec-util
    #   qec-util (pyproject.toml)
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via scalene
kiwisolver==1.4.8
    # via matplotlib
ldpc==2.4.1
    # via dem-decoders
llvmlite==0.46.0
    # via numba
lmfit==1.3.2
    # via
    #   qec-util
    #   qec-util (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.10.3
    # via
    #   pymatching
    #   qec-util
    #   qec-util (pyproject.toml)
    #   sinter
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
networkx==3.5
    # via
    #   pymatching
    #   qec-util
    #   qec-util (pyproject.toml)
numba==0.63.1
    # via galois
numpy==1.26.4
    # via
    #   contourpy
    #   dem-decoders
    #   galois
    #   ldpc
    #   lmfit
    #   matplotlib
    #   numba
    #   pandas
    #   pymatching
    #   qec-util
    #   qec-util (pyproject.toml)
    #   scalene
    #   scipy
    #   sinter
    #   stim
    #   xarray
packaging==24.1
    # via
    #   black
    #   build
    #   matplotlib
    #   pytest
    #   xarray
pandas==2.2.2
    # via xarray
pathspec==0.12.1
    # via black
pillow==11.2.1
    # via matplotlib
pip-tools==7.4.1
    # via qec-util (pyproject.toml)
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
psutil==6.0.0
    # via scalene
pygments==2.18.0
    # via rich
pymatching==2.2.2
    # via
    #   ldpc
    #   qec-util
    #   qec-util (pyproject.toml)
pyparsing==3.2.3
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.3.2
    # via
    #   ldpc
    #   pytest-black
    #   qec-util (pyproject.toml)
pytest-black==0.3.12
    # via qec-util (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-sat==1.8.dev17
    # via qec-util
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   qec-util
    #   qec-util (pyproject.toml)
qec-util[all]==0.3.0
    # via qec-util (pyproject.toml)
rich==13.7.1
    # via scalene
scalene==1.5.44
    # via qec-util (pyproject.toml)
scipy==1.14.0
    # via
    #   dem-decoders
    #   ldpc
    #   lmfit
    #   pymatching
    #   qec-util
    #   qec-util (pyproject.toml)
    #   sinter
sinter==1.15.0
    # via
    #   dem-decoders
    #   ldpc
six==1.16.0
    # via
    #   python-dateutil
    #   python-sat
stim==1.13.0
    # via
    #   dem-decoders
    #   ldpc
    #   qec-util
    #   qec-util (pyproject.toml)
    #   sinter
toml==0.10.2
    # via pytest-black
tqdm==4.67.1
    # via
    #   dem-decoders
    #   ldpc
typing-extensions==4.15.0
    # via galois
tzdata==2024.1
    # via pandas
uncertainties==3.2.2
    # via lmfit
wheel==0.44.0
    # via
    #   pip-tools
    #   scalene
xarray==2024.7.0
    # via
    #   qec-util
    #   qec-util (pyproject.toml)

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
